# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BP Challenge",
        "description": "BP Challenge Swagger OpenApi description",
        "contact": {
            "email": "gbrayhan@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/account": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Get list of persons",
                "description": "Returns list of persons",
                "operationId": "getProjectsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/person": {
            "get": {
                "tags": [
                    "Person"
                ],
                "summary": "Get list of persons",
                "description": "Returns list of persons",
                "operationId": "getProjectsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Person"
                ],
                "summary": "Register Person",
                "description": "Register Person",
                "operationId": "storePerson",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name",
                                        "type": "string",
                                        "example": "Alex"
                                    },
                                    "last_name": {
                                        "description": "last name",
                                        "type": "string",
                                        "example": "Guerrero"
                                    },
                                    "phone": {
                                        "description": "phone",
                                        "type": "string",
                                        "example": "+554422331122"
                                    },
                                    "curp": {
                                        "description": "CURP",
                                        "type": "string",
                                        "example": "GGAA887777HDFBRR01"
                                    },
                                    "address": {
                                        "description": "Address",
                                        "type": "string",
                                        "example": "Av Elementia, Zapopan Jalisco, Mexico"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Person created successfully"
                                        },
                                        "person": {
                                            "$ref": "#/components/schemas/Person"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/person/associate-person": {
            "post": {
                "tags": [
                    "Person"
                ],
                "summary": "Associate Person to the current user",
                "description": "ssociate Person to the current user",
                "operationId": "associatePerson",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "person_id": {
                                        "description": "Person Identifier",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Person successfully associated"
                                        },
                                        "person": {
                                            "$ref": "#/components/schemas/Person"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Person": {
                "required": [
                    "name",
                    "last_name",
                    "phone",
                    "curp",
                    "address"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "Doe"
                    },
                    "phone": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "+52 5544332211"
                    },
                    "curp": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "GGSS923211HDFBRR01"
                    },
                    "address": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Av Elementia 12, Arcos de Vallarta, Guadalajara, Jalisco, Mexico"
                    },
                    "created_at": {
                        "description": "Datetime marker of verification status",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true",
                        "example": "2019-02-25 12:59:20"
                    },
                    "updated_at": {
                        "description": "Datetime marker of verification status",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true",
                        "example": "2019-02-25 12:59:20"
                    },
                    "deleted_at": {
                        "description": "Datetime marker of verification status",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true",
                        "example": "2019-02-25 12:59:20"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Person"
                }
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer [TOKEN])",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}